chore: re-add delegation app files
