+++ components/delegation/NewSubDelegation.tsx
@@ -6,7 +6,7 @@ import { useState } from "react";
 import {
   DelegationCollection,
   SUB_DELEGATION_USE_CASE,
-} from "../../pages/delegation/[...section]";
+} from "./constants";
 import { DELEGATION_CONTRACT, NEVER_DATE } from "../../constants";
 import { DELEGATION_ABI } from "../../abis";
 import { isValidEthAddress } from "../../helpers/Helpers";
